add blazegraph related work for CityGML exporting, especially focus on DBSplitter(1427-dev-export-citygml)

+++ impexp-core/src/main/java/org/citydb/citygml/deleter/concurrent/DBDeleteWorker.java
@@ -130,7 +130,7 @@ public class DBDeleteWorker extends Worker<DBSplittingResult> implements EventHa
 			if (!shouldWork)
 				return;
 
-			long objectId = work.getId();
+			long objectId = (Long)work.getId();
 			long deletedObjectId;
 
 			if (config.getProject().getDeleter().getMode() == DeleteMode.TERMINATE) {
